refactor: domain을 record 형식의 dto로 변경
