+++ src/main/java/KNUCHAT/dto/VideoMessage.java
@@ -1,0 +1,9 @@
+package KNUCHAT.dto;
+
+public record VideoMessage(
+        Long roomId,
+        Long senderId,
+        Long receiverId,
+        VideoCallStatus videoCallStatus
+) {
+}
